Clean test run
