+++ package.json
@@ -22,15 +22,15 @@
         "@ethereumjs/block": "5.2.0",
         "@ethereumjs/common": "4.3.0",
         "@ethereumjs/evm": "3.0.0",
+        "@ethereumjs/rlp": "5.0.2",
         "@ethereumjs/statemanager": "2.3.0",
         "@ethereumjs/tx": "5.3.0",
-        "@ethereumjs/vm": "8.0.0",
         "@ethereumjs/util": "9.0.3",
-        "@ethereumjs/rlp": "5.0.2",
+        "@ethereumjs/vm": "8.0.0",
         "cbor": "^9.0.2",
         "ethereum-cryptography": "2.2.1",
         "fs-extra": "^11.2.0",
-        "solc-typed-ast": "^18.1.7",
+        "solc-typed-ast": "^18.2.0",
         "src-location": "^1.1.0",
         "web3-eth-abi": "^4.2.2"
     },
