+++ src/debug/opcode_interposing.ts
@@ -1,18 +1,5 @@
-import { ERROR } from "@ethereumjs/evm/dist/cjs/exceptions";
-import { RunState } from "@ethereumjs/evm/dist/cjs/interpreter";
-import {
-    OpHandler,
-    Opcode,
-    addresstoBytes,
-    trap,
-    writeCallOutput
-} from "@ethereumjs/evm/dist/cjs/opcodes";
-import {
-    AsyncDynamicGasHandler,
-    SyncDynamicGasHandler
-} from "@ethereumjs/evm/dist/cjs/opcodes/gas";
-import { AddOpcode } from "@ethereumjs/evm/dist/cjs/types";
-import { Address, bigIntToBytes, setLengthLeft } from "@ethereumjs/util/dist/cjs";
+import { Common } from "@ethereumjs/common";
+import { Address, bigIntToBytes, setLengthLeft } from "@ethereumjs/util";
 import { bigEndianBufToBigint } from "../utils";
 import {
     FoundryCheatcodesAddress,
@@ -20,6 +7,28 @@ import {
     RevertMatch,
     returnStateMatchesRevert
 } from "./foundry_cheatcodes";
+
+const EXCEPTION_MOD = require("@ethereumjs/evm/dist/cjs/exceptions");
+const ERROR = EXCEPTION_MOD.ERROR;
+
+const OPCODES_MOD = require("@ethereumjs/evm/dist/cjs/opcodes");
+const addresstoBytes = OPCODES_MOD.addresstoBytes;
+const trap = OPCODES_MOD.trap;
+const writeCallOutput = OPCODES_MOD.writeCallOutput;
+
+/// require(@ethereumjs/evm/dist/cjs/types).AddOpcode
+type AddOpcode = any;
+/// require(@ethereumjs/evm/dist/cjs/types).Opcode
+type Opcode = any;
+/// require(@ethereumjs/evm/dist/cjs/types).OpHandler
+type OpHandler = any;
+
+/// require("@ethereumjs/evm/dist/cjs/opcodes/gas").AsyncDynamicGasHandler
+type AsyncDynamicGasHandler = any;
+/// require("@ethereumjs/evm/dist/cjs/opcodes/gas").SyncDynamicGasHandler
+type SyncDynamicGasHandler = any;
+/// require("@ethereumjs/evm/dist/cjs/interpreter").RunState
+type RunState = any;
 
 function interopseOnOp(code: number, opcodes: any, handler: OpHandler): AddOpcode {
     const originalOp: Opcode = opcodes.opcodes.get(code);
@@ -89,7 +98,7 @@ export function foundryInterposedOps(opcodes: any, foundryCtx: FoundryContext): 
         // 0xf0: CREATE
         [
             0xf0,
-            async function (runState, common) {
+            async function (runState: RunState, common: Common) {
                 // If we're calling vm.* ignore the expectedRevert
                 const expectedRevert = getExpectedRevert(foundryCtx);
                 const [value, offset, length] = runState.stack.popN(3);
@@ -117,7 +126,7 @@ export function foundryInterposedOps(opcodes: any, foundryCtx: FoundryContext): 
         // 0xf5: CREATE2
         [
             0xf5,
-            async function (runState, common) {
+            async function (runState: RunState, common: Common) {
                 const expectedRevert = foundryCtx.getExpectedRevert();
 
                 if (runState.interpreter.isStatic()) {
@@ -178,7 +187,7 @@ export function foundryInterposedOps(opcodes: any, foundryCtx: FoundryContext): 
         // 0x00: STOP
         [
             0x00,
-            function (runState) {
+            function (runState: RunState) {
                 const expectedRevert = getExpectedRevert(foundryCtx);
                 if (expectedRevert === undefined) {
                     trap(ERROR.STOP);
@@ -212,7 +221,7 @@ export function foundryInterposedOps(opcodes: any, foundryCtx: FoundryContext): 
         // 0xf4: DELEGATECALL
         [
             0xf4,
-            async function (runState) {
+            async function (runState: RunState) {
                 const value = runState.interpreter.getCallValue();
                 const [, toAddr, inOffset, inLength, outOffset, outLength] = runState.stack.popN(6);
                 const toAddress = new Address(addresstoBytes(toAddr));
@@ -239,7 +248,7 @@ export function foundryInterposedOps(opcodes: any, foundryCtx: FoundryContext): 
         // 0xfa: STATICCALL
         [
             0xfa,
-            async function (runState) {
+            async function (runState: RunState) {
                 const value = BigInt(0);
                 const [, toAddr, inOffset, inLength, outOffset, outLength] = runState.stack.popN(6);
                 const toAddress = new Address(addresstoBytes(toAddr));
@@ -261,7 +270,7 @@ export function foundryInterposedOps(opcodes: any, foundryCtx: FoundryContext): 
         // 0xf3: RETURN
         [
             0xf3,
-            function (runState) {
+            function (runState: RunState) {
                 const [offset, length] = runState.stack.popN(2);
                 let returnData = new Uint8Array(0);
